# api-docs/openapi.yaml
openapi: "3.0.0"

info:
  title: "My API documentation"
  version: "1.0"
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: /api
    description: Base Path for all API calls
    variables: {}

paths:
  
  /query:
    get:
      operationId: run.query_metric
      parameters:
        - name: sensor_ids
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
        - name: metrics
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
        - name: statistic
          in: query
          required: false
          schema:
            type: string
            enum: [min, max, sum, average]
            default: average
        - name: date_range
          in: query
          required: false
          schema:
            type: string
            description: Date range format should be 'YYYY-MM-DD:YYYY-MM-DD'. If not specified, defaults to 'latest' (last 24 hours).
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  sensors:
                    type: object
                    additionalProperties:
                      oneOf:
                        - type: object
                          properties:
                            statistic:
                              type: string
                            value:
                              type: number
                        - type: object
                          properties:
                            value:
                              type: number

  /metric:
    post:
      operationId: run.add_metric
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sensor_id:
                  type: string
                metric:
                  type: string
                value:
                  type: number
                unit_of_measurement:
                  type: string
                timestamp:
                  type: string
                  format: date-time
      responses:
        201:
          description: Data added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /greeting/{name}:
    post:
      operationId: run.post_greeting
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Successful response
          content:
            text/plain:
              schema:
                type: string